First update 7/12/2017
Updated flakes_game_template.html. Added functions for the basic
equations and displayed them on the flakes game page. Including a label
for answer for testing purposes.

Also included a bit of drawing with stroke on the canvas. Later plan to
use this to create a drawing with the values of the problems x or x and
y values.